+++ src/sider/oppfolgingsplan/oppfolgingsplaner/BeOmOppfolgingsplan.tsx
@@ -1,19 +1,35 @@
-import { BodyLong, BodyShort, Box, Button, Heading } from "@navikt/ds-react";
+import {
+  Alert,
+  BodyLong,
+  BodyShort,
+  Box,
+  Button,
+  Heading,
+} from "@navikt/ds-react";
 import React from "react";
 import { useValgtPersonident } from "@/hooks/useValgtBruker";
 import { NarmesteLederRelasjonDTO } from "@/data/leder/ledereTypes";
 import * as Amplitude from "@/utils/amplitude";
 import { EventType } from "@/utils/amplitude";
-import { NewOppfolgingsplanForesporselDTO } from "@/data/oppfolgingsplan/oppfolgingsplanForesporselHooks";
-import { UseMutationResult } from "@tanstack/react-query";
+import {
+  NewOppfolgingsplanForesporselDTO,
+  useGetOppfolgingsplanForesporselQuery,
+  usePostOppfolgingsplanForesporsel,
+} from "@/data/oppfolgingsplan/oppfolgingsplanForesporselHooks";
+import { tilLesbarDatoMedArUtenManedNavn } from "@/utils/datoUtils";
+import { OppfolgingstilfelleDTO } from "@/data/oppfolgingstilfelle/person/types/OppfolgingstilfellePersonDTO";
 
 const texts = {
+  aktivForesporsel:
+    "Obs! Det ble bedt om oppfølgingsplan fra denne arbeidsgiveren",
   header: "Be om oppfølgingsplan fra arbeidsgiver",
   description:
     "Her kan du be om oppfølgingsplan fra arbeidsgiver eller purre om det mangler.",
   virksomhet: "Virksomhet:",
   narmesteLeder: "Nærmeste leder:",
   button: "Be om oppfølgingsplan",
+  foresporselSendt: "Forespørsel om oppfølgingsplan ble sendt",
+  foresporselFeilet: "Det skjedde en uventet feil. Vennligst prøv igjen senere",
 };
 
 function logOppfolgingsplanForesporselEvent() {
@@ -28,19 +44,29 @@ function logOppfolgingsplanForesporselEvent() {
 
 interface Props {
   aktivNarmesteLeder: NarmesteLederRelasjonDTO;
-  postOppfolgingsplanForesporsel: UseMutationResult<
-    NewOppfolgingsplanForesporselDTO,
-    Error,
-    NewOppfolgingsplanForesporselDTO,
-    unknown
-  >;
+  currentOppfolgingstilfelle: OppfolgingstilfelleDTO;
 }
 
 export default function BeOmOppfolgingsplan({
   aktivNarmesteLeder,
-  postOppfolgingsplanForesporsel,
+  currentOppfolgingstilfelle,
 }: Props) {
   const personident = useValgtPersonident();
+  const getOppfolgingsplanForesporsel = useGetOppfolgingsplanForesporselQuery();
+  const postOppfolgingsplanForesporsel = usePostOppfolgingsplanForesporsel();
+
+  const lastForesporselCreatedAt =
+    getOppfolgingsplanForesporsel.data?.[0]?.createdAt;
+  const isAktivForesporsel =
+    !!lastForesporselCreatedAt &&
+    !!currentOppfolgingstilfelle &&
+    !postOppfolgingsplanForesporsel.isSuccess
+      ? currentOppfolgingstilfelle.start <= lastForesporselCreatedAt &&
+        lastForesporselCreatedAt <= currentOppfolgingstilfelle.end
+      : false;
+  const aktivForesporselTekst = `${
+    texts.aktivForesporsel
+  } ${tilLesbarDatoMedArUtenManedNavn(lastForesporselCreatedAt)}`;
 
   function onClick() {
     const foresporsel: NewOppfolgingsplanForesporselDTO = {
@@ -56,25 +82,44 @@ export default function BeOmOppfolgingsplan({
   }
 
   return (
-    <Box background="surface-default" className="mb-4 flex flex-col p-4">
-      <Heading size="small" level="3">
-        {texts.header}
-      </Heading>
-      <BodyLong className="mb-2">{texts.description}</BodyLong>
-      <BodyShort>
-        {texts.virksomhet} {aktivNarmesteLeder.virksomhetsnavn}
-      </BodyShort>
-      <BodyShort className="mb-2">
-        {texts.narmesteLeder} {aktivNarmesteLeder.narmesteLederNavn}
-      </BodyShort>
-      <Button
-        className="w-fit mb-2"
-        size="small"
-        onClick={onClick}
-        loading={postOppfolgingsplanForesporsel.isPending}
-      >
-        {texts.button}
-      </Button>
-    </Box>
+    <>
+      {isAktivForesporsel && (
+        <Alert variant="warning" className="mb-2">
+          {aktivForesporselTekst}
+        </Alert>
+      )}
+      <Box background="surface-default" className="mb-4 flex flex-col p-4">
+        <Heading size="small" level="3">
+          {texts.header}
+        </Heading>
+        <BodyLong className="mb-2">{texts.description}</BodyLong>
+        <BodyShort>
+          {texts.virksomhet} {aktivNarmesteLeder.virksomhetsnavn}
+        </BodyShort>
+        <BodyShort className="mb-2">
+          {texts.narmesteLeder} {aktivNarmesteLeder.narmesteLederNavn}
+        </BodyShort>
+        {!postOppfolgingsplanForesporsel.isSuccess && (
+          <Button
+            className="w-fit mb-2"
+            size="small"
+            onClick={onClick}
+            loading={postOppfolgingsplanForesporsel.isPending}
+          >
+            {texts.button}
+          </Button>
+        )}
+        {postOppfolgingsplanForesporsel.isSuccess && (
+          <Alert inline variant="success">
+            {texts.foresporselSendt}
+          </Alert>
+        )}
+        {postOppfolgingsplanForesporsel.isError && (
+          <Alert inline variant="error">
+            {texts.foresporselFeilet}
+          </Alert>
+        )}
+      </Box>
+    </>
   );
 }
